requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.23
jinja2>=3.1
pyyaml>=6.0
aiofiles>=23.0
